[FIX] l10n_mx_edi: Fixed assignation to l10n_mx_edi_origin
Before of this fix We get error in the next cases:

** CASE 1

1.- Generate a customer invoice
2.- Validate the invoice
3.- Cancel the invoice generated
4.- Reset2draft the invoice
5.- Validate again the invoice
6.- Try generate an invoice refund to the invoice.

** CASE 2

1.- Generate a customer invoice
2.- Validate the invoice
3.- Generate a invoice refund
4.- Validate the invoice refund
5.- Try cancel the invoice refund

This because in the last step to each case the l10n_mx_edi_origin type
is different to the value that already have the record.

In this cases, not must be generated an raise, must be changed the
l10n_mx_edi_origin type by the new type.