[FIX] web_studio: forbid to drop group inside group
We don't want to allow the user to drop columns inside a column in the form view.
Because this breaks the design and is not user friendly.
Furthermore, the tests are fixed since it was impossible to chain two drag and
drop due to the throttle on the drag event.