[FIX] web_gantt: crash when no data
- Install `hr_holidays` without demo data
- Access the App

It crashes because `task.task_stop.getTime()` is not defined.

When no data is available, we use `self.focus_date` to fill in a dummy
group. However, `self.focus_date` is a `Moment` object, not a `Date`
object. On a `Moment` object, `getTime` is not defined.

We could have used `valueOf`, which works for both objects. But better
be consistent :-)

opw-775900
opw-776065