[FIX] quality_mrp: prevent quantity/lot/SN changes from work orders 'manager view'
This revision ensures that users cannot modify the current quantity or
enter lot numbers from the so-called manager view. This is because the
'tablet view' is not aware of the changes made in the manager view.

This was previously done in community (see
https://github.com/odoo/odoo/commit/44f06a66467cb39780fa65d0cefb44704b072e94).
The decision to do it in quality_mrp is motivated by the desire not to
affect users who still want to use work orders without quality.