[FIX] website_crm_score: track routes based on their viewid in dom
Since every routes have a template (even if they don't return an ir.ui.view
as main object), we use it to store the track state of the page.

This way, the behavior/use case is the same for static or dynamic pages.

This commit closes #1525
This commit depends of e19c6a5