[FIX] website_crm_score: allow routes to be trackable
Before this commit, visiting a route set as 'track=True' in controller would
not track page since the super was call after the code doing the track stuff.
It would also delete the cookie 'lead_id'.

Now, it is correctly tracking controllers routes which have 'track=True',
and the cookie wont be deleted