[FIX] web_studio: image field in list view before new views
Studio 10.0 uses a preview version of the new views introduced in
saas-16. In these new views, image fields in list views with the
`widget="image"` attribute are displayed as images. In the old views,
this was not the case, a download link was displayed instead (as
'file' fields).

This commit makes the old new views used by Studio display the list
image fields as binary fields.