[FIX] account_reports: cash_basis option wasn't processed correctly.
OPW 772107.
Use case: open the general ledger from the cash flow statement, the amounts were not equal. The reason was simply that the option cash_basis was't set in the context for being used in get_with_statement(). Now it's the case, although the users have still to manually thick that option when coming from the cashflow (which is a bit sad, but hard to solve with current design)