[FIX] web_studio: allow stat buttons with same m2o field name
Before this commit, it was in some case not really possible to add two
stat buttons in the same form view.  More precisely, if we select two
many2ones (in the comodel) with the same name (in the comodel), there
was a name conflict.

The way it worked is that we create a computed field on the model, which
compute the number of occurences in the relation.  In this commit, we
simply choose a field name which also depends on the model name, to make
sure no conflict can occur again.