[FIX] web_studio: swap replace xpaths
When deleting an existing element from a view (i.e. replace xpath),
one might want to add the same existing element somewhere else in
the view (effectively emulating moving the element), if the new
position happened to be before the original position, the element
at the new position would be the one being replaced since the
replace xpaths are being appended to the arch.

To fix this, a new function add_and_swap has been added inside the
body of the normalize method, this function will prepend replace
xpaths to the arch instead of appending them.