[IMP] web_studio: binary fields should keep the file name
Previously, downloading a file from a Studio-created binary field
yielded a file whose name was that of the associated field, without any
extension. This caused a serious usability issue, as it was the user's
responsibility to select the right software to open the file.

This commit ensures that a binary field is always dropped with an
associated char field (whose role is to contain the file name with its
correct extension). As a result, files uploaded via a Studio binary
field keep their names when they are subsequently downloaded.