[FIX] web_studio: properly anchor template fields
Before this commit, the normalization algorithm would
anchor template elements directly inside the kanban element,
which wouldn't make it show up in the actual interface
because in order for it to show up in the interface it
has to be inside the templates element.

This is fixed by checking if the template element is an
ancestor of the current element being examined during
the xpath expr generation, and appending "templates//"
to the generated expr.