[FIX] web_studio: correctly set default view
Setting a view as 'Default' in Studio sometimes didn't work if the field
`view_id` was set on the action.

There are three different ways to define the views order (in order of precedence):
 * `view_ids`
 * `view_id`
 * `view_mode`

So if `view_id` is defined, setting the view_mode won't have any impact.
This special case wasn't correctly managed before this rev.