[IMP] studio: app switcher reset default background
On the App Switcher (Home Menu), when Studio is installed,
the user is now able to set back the background to default easily.

A new menu item 'Reset Default Background' has been added
to the top-left side menu, just below 'Change Background'.

In order to avoid misclick from the user,
a dialog asks him or her to confirm the action.