[IMP] website_crm_score: simplify tracking route.
track=True on route is no more supported.
Now we use the mechanism on view directly.

We keep the template name when we render the page and reuse it in the dispatch
to know if this view need to be tracked or not.

That's make it configurable by end user without any python code.

This commit depends of e19c6a5