[FIX] web_studio: set studio_groups attr after RNG validation
Since RNG validation was restored, editing views that contain group
constraints has been impossible. As an example, fields from the tree
view of 'purchase order' cannot be renamed via Studio. This is due to
the fact that the groups associated with a field are given to the
web_client via the RNG-invalid 'studio_groups' attribute. In particular,
any attempt to edit such a view will result in the user being prompted
with a view validation error.

We fix the issue by ensuring that the RNG validation is performed before
the studio_groups attribute is set.