[FIX] web_studio: navigation with new views
With the new views, a small change has been introduced but it has impacted the
navigation in Studio. Before the new views, `do_action` was either managed by
the ActionManager (if the widget doing the `do_action` was a child of the AM) or
the WebClient (otherwise). Now, `do_action` uses the trigger_up mechanism and is
always handled by the WebClient.

This change breaks some complex navigation in Studio (like importing automated
action through Studio) but also allows to simplify the navigation.

As everything goes through the WebClient `do_action`, the actions don't need
to manually specify themselves the `keep_state` and `disable_edition` attribute
anymore.

Actions can be distinguished in two types with Studio:

* the 'navigation' action, i.e. the ones which will change the Studio environment
  and edited ; for this, they are launched in another action manager. The URL
  state changes for these actions. Note that these actions can be easily
  identified: the ones linked to menu (and application) clicks.
* the rest, i.e. actions inside Studio that need to be handled normally but
  the URL state can't change as the action edited by Studio doesn't change

This differenciation can thus directly be done in the WebClient from now on.

Note that a small refactoring was necessary in the enterprise web client to
allow overriding the options passed to `do_action` when clicking on menus.