[FIX] account_reports: language mail customer
In 45200620 the customer statements report was modified so it is
consistent with the "Overdue" button on a customer.

Thanks to this, we now see at both locations a customer statement in
the partner language, but there was still an issue: if when sending the
mail, the mail had not be changed, the original text would be gotten but
in the current user language because it was overrided at one point.

With this fix, we avoid overriding the summary if possible when getting
the report informations.

opw-766594
closes #1554