[FIX] web_studio: disable events on widgets in form editor
When clicking in Studio on a widget with bound events (many2one for example),
we expect to edit its field properties (and not open the form view in the
previous case).

To prevent the original handlers to be executed, all events are removed from
the widgets and only the Studio events (click, hover, etc.) remain.