[FIX] website_sign: remove wrong label in xml view
In the form view definition of the signature.request.template model,
there were a label for the "name" field... but there is no "name" field
in that model. Before the new views, this was not breaking anything, the
label was just not shown. In the new views, this throws a traceback.