[FIX] web_enterprise: app switcher search with japanese ime (#1558)
BUG 1:
With a Japanese IME, whenever there are matching menu
items, the input text is no longer in composition mode
(i.e. underlined).

REASON:
Whenever there are matching items, the input loses focus.
It should never be the case during composition mode!

FIX:
Remove a change of focus when we are in composition mode.

=========================================================

BUG 2:
When the focus is not on the search input, the first
keyboard event will never be in composition mode.

REASON:
The composition mode may become active only if
the target of the keydown event is editable.

(PARTIAL) FIX:
At page loading, the focus is on the hidden search bar.

========================================================

REMAINING BUG:
If the focus is not on the search bar, the next
keyboard event will never be in composition mode.

WORKAROUND:
Press ESC prior to typing a composing text.
ESC forces the focus on the search bar.