[FIX] website_version: don't try to render a view with ID False
This occurs when having one QWeb view (e.g. website.homepage) with a
version_id and attempting to render that without a version.

This leads to the super render being called with False as it's first
argument which leads to ir.ui.view's read_combined being called with
self being False. In there the following loop will never end:

root = self
while root.mode != 'primary':
    ...
    check_view_ids.append(root.id)
    root = root.inherit_id

This avoids falling back to a website_version specific view when it
doesn't exist.

Note that this didn't happen before
9131d3193b2f9d2952517b008a9a0cb1b9c761c4.

opw-779334