[FIX] web_mobile: datepicker: value not correctly set
In mobile, we want to display the mobile native date picker
but there was an error when selecting a value.

We reproduce here the behavior of the datepicker by directly
set the value chosen in the input field.

In addition, we have to trigger the 'datetime_changed' event
to notify the changes.
'commit_value' doesn't exist anymore and was replaced by
'commitChanges' but it's not necessary to do this in this case.