[FIX] web_studio: add a one2many related field
When adding a one2many related field with Studio, the attributes
`relation` and `relation_field` were not correctly set.

These two attributes represent the m2o field which the o2m points on.
This piece of information is not known by the FieldSelector (only the o2m field
name and model are), so we need to set it when creating the field with Studio.

See opw-778672