[FIX] website_quote_subscription: subscription in portal
- Create a product A with 'Subscription Product' set
- Create a SO, add A and confim.
- In the portal, go to 'My Account > Sales Orders', and open the create
  SO.

The subscription line don't appear.

Using `_split_subscription_lines` is not appropriate to render the
template. This method is intended to be used in order to know which SO
line should trigger the creation of a subscription. Therefore, it
filters out subscription lines from the result.

We create a specific `_get_subscription_lines` method which sole purpose
is to handle the template rendering. This covers subscription products
with or without a linked Subscription Template.

opw-777742