[FIX] web_mobile: datepicker issues on iOS and Android
(1) Avoid datepicker crash on iOS:
Before this commit, it was not possible to open a datetime field
on iOS. The application was killed.
self.getValue() returns a moment js object but we want to be sure
of the date format sent to the mobile app.

(2) Ignore timezone issue on Android:
Before this fix, the time was sent in UTC to the app.
Due to a mismatch with the webview and the app settings, the
offset was always the same. So, for summer dates, the current
offset (GMT+1 Brussels) was applied instead of GMT+2.
To make it more simple and to avoid useless convertion, we decided
to directly send the local time to the app without having
to convert it again.
So, we use the local time until the save on the server.
A fix in the mobile repository will follow to handle the local
time instead of the UTC time to convert.