[FIX] web_studio: generalize subview and kanban anchor
Previous to this commit, if one would add an existing field to a subview
whose anchor element already existed in the parent view, the new field
would anchor to the element in the parent view and not to the one in the
subview.

This happened because, when calculating relative xpaths, we go off on
the assumption that there can only exist one instance of a named field
within a view, which is false, and this was patched for kanban views by
prepending //templates to the xpath generated if the anchor element was
within a kanban view.

This is fixed by forcing an absolute path calculation if the count of
the elements with the anchor element's name is more than one, instead of
using a relative xpath based on name (which would match the first
element found).

With this, the code that handled kanban view cases can be removed since
it's a generalization.

opw-778051