[IMP] web_grid: supports 'week' as step
Grid view now supports to display a month (span)
by week (step). This implementation chosen is to
display a year week (from monday to sunday, following
ISO-8601).

We choose to display week such as
- Some days of the week can't belong
to the displayed month. Example: Week 31 of 2017
starts on 31 july, and ends with the 6 august.
- On pager, clicking on 'next' can make visible
the same last week (column) as the first one of
the new page, as might overlap 2 months.
- Displaying August 2017 by month will show days
between 31 july and 3 sept.
- The name of column in grid view are the number
of the week and its first and last days
(e.i.: "Week 32, 5 July - 13 July").