[IMP] project_forecast: introduces time unit and span
This commit introduces the time unit and time span
for forecast, per company:
- time unit: you can choose to forecast by hours or
by days. Each number introduced on grid view will
be expressed in that company unit.
- time span: you can choose to display and encode
your forecast by period (day, week or month). The
'step' of the grid view will be the span from the
company.

This makes forecast usable since it covers the case
of accurate forecast (precision in hours) and more
global resource allocation (allocate days in month).

To avoid depending on master data, the conversion
factor 'hour/day' does not depend on the product.uom
data but on the average working hour (per day) from
the resource.calendar of the project or the company.