[FIX] web_gantt: readonly mode drag-and-drop
Before this fix, it was possible to change the start and end dates
of a task with drag-n-drop, even when edit=false.

With this fix, the drag-n-drop interactions on tasks are disabled
when edit=false.

Also, since the object 'gantt' of the library dhtxml is shared among
all gantt views, we fixed a bug where some config attributes were not
refreshed at rendering (e.g. 'round_dnd_dates').