[FIX] web_studio: check for related monetary field
In a previous revision (see odoo/enterprise@a67c889) we have added a sanity
check when adding a monetary field ; you can't add a monetary field if there
is no `currency_id` field on the model.

This check was not done if the field was a related one. This is now done.

Note that `x_currency_id` is also a comptatible field (see odoo/odoo@4f62a65) so
the condition has been also adapted.

opw-778578