[IMP] web_studio: rename fields added with Studio
This commit introduces a way to rename the technical name of a field freshly
added with Studio.

When creating a new field with Studio (drag&drop), a name is randomly built
(x_studio_field_[6 random characters]). This random name was bothering users
that needed to use it when creating domain or server actions for example.

We thus introduce the possibility to rename this technical name: when clicking
on a field, a new section "Technical Name" appears in the sidebar (in debug mode),
which can be modified if the field has just been created with Studio, in the
same edition session.

Moreover, for non technical users, the technical name is also automaticcaly
modified if the field label is modified during the same edition session as the
field addition ; if the user is renaming a field label "Number of Shoes", the
technical field name will be "x_studio_number_of_shoes".