[FIX] web_studio: click on a page in group
Note: there is no need to try to forward-port this in saas-16 as this situation
has been prevented.

It is possible to add a notebook in a `<group>` with Studio.

Before this rev. clicking on a page of a notebook in a group was selecting
the group instead of the page.

An attempt has been previsouly done to fix it (with a `stopPropagation`, see
odoo/enterprise@66d0f44) but reverted afterwards because it prevents the page
to be opened (see odoo/enterprise@c8cf921).

The page is now manually opened in JS so that the `stopPropagation` can still
be used.