[FIX] web_studio: add calendar or gantt view
When adding a new calendar or gantt view with Studio, a popup appears for
the user to choose some fields for mandatory view attributes (`date_start`
for example).

Before this rev. the error message was used to determine which view couldn't
be created ; this doesn't make any sense as the view type is already known and
the error message can be translated.

The mentioned view types couldn't be added in Spanish for example.

opw-779010