[FIX] l10n_be_coda: globalisation
- Import a CODA file containing 2 globalized movements:
- The first movement is not detailed
  Amount: 100
- The second movement is detailed
  Amount: 200
  Detail: 190
            5
            3
            2

Odoo imports 100, 200, 190, 5 and 3. It is incorrect, it should import
100, 190, 5, 3 and 2.

The problem lies in the globalisation stack used. We use character 125
of the line to determine is the movement is globalized, but also to keep
track of the beginning/end of the details. This character is indeed used
to define the bginning and the end of a globalisation. If the character
used for the two movements is the same, the beginning of the second
movement will be considered as the end of the first movement. This is
incorrect.

We build the globalization stack using the movement reference instead,
to get a unique number.

opw-781684