[FIX] web_studio: edit label string in form
The rev. odoo/enterprise@c9e331a introduced a new handler to edit label string
in form (like <label for='...' string='...'/> tags).

The handler wasn't restrictive enough and all labels were clickable (including
the ones associated to fields), which wasn't desired. The handler is now removed
for field labels.

This was causing a traceback when clicking on a o2m or m2m label in Studio.

See opw-783197