[FIX] web_studio: correctly export act_window.view
When exporting customizations in Studio, if the user had created some
`ir.actions.act_window.view` records (for example by adding a list view
in Project), this module couldn't be imported later on.

This was due to the fact that this specific model wasn't exported.

Note that the one2many `view_ids` on `ir.actions.act_window` has been removed
because it will be filled when setting corresponding many2one on the added
model.

opw-783495