[FIX] project_timesheet_synchro: imported records not synchronised
Before this commit, some projects, taks or AAL* didn't appear in the app.
This is due to a non correctly formatted XML ids.
The app didn't find any record related to this XML id on the server and
the record was ignored.

When you import a record in Odoo, you can specify an id without prefixing it
by the name of the module. e.g. 'project_project_x'
The result of this is an empty string as module name.
Maybe we should fix this in master.

The problem was that the app tried to search on '.project_project_x'.
This is the result of: module + '.' + name

Now, we ensure that the xml id contains a dot even if the module name
is an empty string.

*Analytic Account Line

opw-781410