[FIX] web_studio: create one2many w/o warning
When creating a new one2many field with Studio, a server warning was
displayed because the `ir.model.fields` record was created with an
unkown value `relation_field_id`.

This value comes from the JS and is used to set other o2m values (like
`relation` and `relation_field`) but shouldn't be passed at creation.