[FIX] web_studio: export module on Windows
When generating the `data` attribute in the `__manifest__.py` file
of the exported module, Studio uses the function `os.path.join`.

If the module is exported in Windows, this will result in filepath like
data\\ir_model.xml instead of data/ir.model.xml

The filepath thus needs to be joined with `/`.

opw-783250 opw-781991