[IMP] core: menu tip ux/ui improvement
We would like to make the "no item found" screens more appealing.

Before this commit, it shows a small help tip in the top-left
corner of the screen, just below the "Create" button.

With this commit, these help tips have been replaced by onboarding
screens, which consist of a picture and some text below, both of which
are horizontally centered.

The texts have been slightly changed, so that they are shorter and clearer.

Considered modules:

(A)
    account_batch_deposit,
    account_deferred_revenue,
    account_online_sync,
    account_reports,
    account_sepa_direct_debit
(H)
    helpdesk,
    hr_appraisal,
    hr_contract_salary
(L)
    l10n_mx_edi
(M)
    marketing_automation,
    mrp_plm,
    mrp_workorder
(P)
    pos_loyalty
(Q)
    quality,
    quality_control,
    quality_mrp
(S)
    sale_coupon,
    sale_subscription
(V)
    voip
(W)
    web_enterprise,
    web_grid,
    web_studio,
    website_calendar,
    website_crm_score,
    website_sign