[FIX] stock_barcode: on_barcode_scanned
This method is intended to be called in an onchange context, when called
in a write context the split of move lines done by some of the _check*
methods did not work as expected and broke the reservation. Obviously,
the _check* methods should not do anything about the reservation but
anyway they're all designed to work in an onchange's context (use of
.new, etc). And the whole barcode logic should only uses onchange.

To get rid of the write's usage, we get rid of the javascript calling
the method after a save. This was done to be able to scan a non reserved
product tracked by lot and opens the lot wizard directly, but the sides
effect are too high and we go back to the behavior of v9 and v10: if the
move line was not present, you have to scan one time to make it appear
and a second time to open the lot's wizard.

This method is kept as it's public and this patch is made on a stable
realse, but it isn't called nor used anymore.