[REF] web_studio: put static files in a separate bundle
As Studio is only available for `group_system` users, it is not useful to load
Studio code for other users.

Before this rev. the verification was done in the JS files, with the key
`is_system` present in the session but this wasn't very safe, convenient and
optimized.

The JS and LESS files for Studio are now placed in a new assets bundle only
loaded for the allowed users.

Note that some less mixin (.container and .text-overflow) are not available in
an isolated bundle so they have been replaced by an equivalent.

I also took this opportunity to clean useless `requires` in Studio JS files.