[FIX] web_enterprise: navbar style when toggling home menu
Before this rev., the navbar style wasn't correctly synchronized
with the home menu (the navbar is supposed to look differently when
the home menu is toggled). In particular, open an record in edit
mode, change something, toggle the home menu: a confirmation dialog
opens, to ask if the changes should be discarded, but the navbar
style is already changed. So if the user clicks on 'Cancel', it is
desynchronized.