[REF] voip: use category activity type instead of new field
voip introduces a new field in order to tag some activity types as
being phonecalls. However a field is already present on the activity
type model to implement some custom behavior on activity types.

This commit fixes the voip code to use the category field instead of a
custom one. This should have been done directly when merging voip as it
was part of the original technical spec. However code review was not
performed completely, meaning this modification is done only now.