[IMP] hr_appraisal: improve activity management on appraisal model
This commit improves appraisal management through a better integration of
activities and addition of automated activities. Several things are done in
this commit :

 * automatic activities generation is added for appraisal interview meetings
   and about surveys to fill;
 * a menu to configure activity types is added. Indeed HR managers should be
   able to see and configure activity types related to their job;

Having automated activities allow to replace some messages and tracking
that were implemented to warn people of new appraisals, new meetings or
new surveys to fill. This commit therefore

 * simplifies the tracking as no specific tracking is done anymore;
 * removes to new, scheduled and sent subtypes. This allows to simplify
   chatter a lot in hr_appraisal app;