[IMP] project_timesheet_synchro: LS records prefixed by '__import__'
The Awsome Timesheet app has a bidirectional synchronisation based on
the import/export feature of Odoo.

So, some records created in the local storage need to be imported
in the database.

Before this commit, the xml ids of these record were prefixed
by '__export__' to avoid to have an empty string (*).
But those are imported records, not exported records!
As the default module is now '__import__' (*),
we need to be coherent and change the prefixed module name
of the xml id.

* see related commit odoo/odoo@bb1602b3ac