[REF] quality: split in quality and quality_control
mrp_workorder steps has been build upper quality models.
Currently if the user want to use steps, he has to install
quality.

The purpose of this commit is to split the quality module in
two parts:
- First part with all the common code shared between quality and
workorder steps. It's not an app but a module that will be install
either by quality app or by workorder app.

- Second is quality control. It's the quality module that can be
install by the user. It should contains all the view and code used
exclusively by the quality app.

This commit only moves files and adpat manifest in order to keep a
git history.