[FIX] stock_barcode: inventory: _get_records
In the simplest scenario, the user works with a single location and scan
the product. The javascript mechansim tries to get the line in the
existing ones and if it finds fine, it'll increase its quantity and
prevent the call to onchanges, which could be heavy with numerous lines.

The condition to find back an existing line was wrong because it checked
if an existing line has a product with the barcode just scanned and if
the location is the value of the field `scan_location_id`. In the
simplest scenario, the line has already a location_id set to the main
location while `scan_location_id` is empty, so a line was never found
and the onchanges were always called.

`scan_location_id` allow to work with sublocations (scan the new
location and then scan the products).

opw 767538