[FIX] helpdesk: do not display 'View' button
When adding a partner (without any backend or portal access) as a
follower of an helpdesk ticket, the mail received contains
a 'View ...' button. When clicked, it redirects to the login page.

This feature is not yet complete in v11. So simply hide the button until
this is properly developed in master.

opw-785824