[FIX] mrp_account: cost structure report: produce more
When you product more than the initial demand of a manufacturing order,
the initial demand is not updated after the manufacturing is done. This
leads to inconsistencies in the cost structure report where the cost of
raw material is correctly computed across all consumed materials, but
the cost per product will divide this cost of raw material by the
initial demand of the mo, leading to a cost per product too high in case
of over-production.

opw 786942