[FIX] sale_contract,website_contract: stop renaming an existing AA
When an order that creates a subscription is confirmed, it will either
create a new analytic account along with the subscriptions or use the
one already set on the quote. In that case, it should not rename the
analytic account to avoid confusion.

This commit also add the same auto_commit context mechanism detection as
introduced in 485471e, and some tests for this use case to avoid
regressions.