[FIX] account_sepa: SEPA Credit Transfer XML now triggers
Previous to this fix, the server action default "Action To Do"  (technical name: state) is "Update the record". As the action is meant to be executed as python code, nothing will happen.
By changing the Action to do to "Execute Python Code", the action now triggers from the account.payment tree view, allowing to generate the file and display the download wizard.