[FIX] voip: feedback method should work in batch
voip inherits from mail.activity model to implement behavior linked
to phonecall activities. It inherits from the action_feedback method.
Original method has been written to allow working in batch. Override
does not respect that behavior. This commit fixes the inherit to
effectively work in batch.

This is a manual backport of a saas 11.2 commit see 8bcd90c165cabfa10e7db5b383df5ab880d6f513 .
It is backported even if it is not a real fix because it causes issues
with server action and automated actions. Those could have been done
on activities working in batch which is not the case when voip is
installed.

Done after feedback from internal migration.