[FIX] website_calendar: all_day typo tz/no tz
There was a comparison of date with or without timezone which caused an
error.

Also `.replace(tzinfo=...)` does not do what is expected some times, so
convert it as shown in http://pytz.sourceforge.net

And thirdly an all_day of day 1 to day 2 starts at day 1 00:00:00.00 and ends
at day 2 23:59:59.

opw-788144
closes #1711