[FIX] account_accountant, account_reports: menu
When upgrading from Community to Enterprise, an error occurs because of
an incorrect dependency.

`menu_account_financial_reports_tree` should not be defined in
`account_accountant` since the auto-install of `account_reports` is not
triggered during the upgrade.

We define it directly in `account_reports`. It is moreover more logical
since the action is defined in `account_reports`.

opw-785563