[IMP] project_timesheet_forecast_sale: revamp project overview
Adaptation of revamping project overview: it now
depends on project and not AAL.

The timesheet/forecast table shows timesheet and
forecast even if they are not linked to SO lines.
Two columns are added:
- "before" shows the timesheet before the 3 months displayed
- "after" shows the forecasts after the 3 months displayed
The timesheet part of the table can be seen when 'allow forecast'
is uncheck on project. Some cosmetic changes are also done.

The code is now simpler, using SQL query directly to get
forecast and timesheet raw data, instead of using the ORM.