[FIX] quality_mrp: don't assume at least one move
Before this commit, in _compute_component_id we used to assume that
there would be at least one stock.move whose state wasn't done or
cancelled, and after filtering those 2 states, we would get the first
element in the recordset.

However it is not the case when the work order is in sate done or
cancel, and the computation of the field 'test_type' has no sense when
the workorder is done or cancel.

This is fixed by filtering the workorder on their state before computing
the field.

OPW 801622