[FIX] website_sign: translation at signing
As admin:
- Create a user with language different from English, e.g. French
- Create a document to sign, send it to the user

As the user:
- Login in Odoo
- Click on the link in the received email

The page is partially translated. The top of the page (server template)
is correctly translated in French, while the core of the document (tips,
modal for signature...) are not translated.

The tips can simply be made translatable at the level of the model. The
rest of the page is more tricky: the translations should be handled by
the JS layer, but unfortunately no mechanism is foreseen for such a use
case. We are neither in the backend nor in the website (`website_sign`
doesn't depend explicitly on `website`). The translations are simply not
loaded, so we need to manually load them.

opw-786552