[IMP] tests: adapt enterprise tests to the tests tags system
As the possibility to test tags was introduced, some tests must me
adapated to work as before.
Tests that have to run at 'post_install' have to be tagged '-at_install'
and 'post_install'.
Also, nightly tests that contact external entities have to be tagged
'-standard' to avoid to be launched by default. Also they have to be
tagged 'external' so that they can be selected alone.