[FIX] l10n_mx_edi: fix payment_policy in CFDI
Since the payment term number of days could be specified towards the end of the month, we need to check if the invoice.date_due and the invoice.date are the same to set the payment_policy correctly (if invoice.date == invoice.date_due then payment policy is PUE, else PPD). 

That triggers the add of the payment complement in the CFDI as well.

Was PR #1571. Courtesy of Vauxoo