[FIX] marketing_automation: Allows to post replies on partners.
Purpose
=======

On a mass mailing we can choose to either reply to an address (an alias is mandatory, otherwise it won't work) or to the specific record. This as been restricted for the res.partner. It doesn't make sense as this record has a chatter.

Specification
=============

Correct that. Also fix the glitch with the fields position.