[IMP] helpdesk: handle access token in helpdesk mail/view controller
This commit adds support in helpdesk of access token used in mail/view
controller. Indeed mail propagates the access token given to the
controller. Checking its content and redirecting users to the right
URL is a behavior that belong to the application. This commit adds
the support of access token in helpdesk, like what already exists for
sale orders and account invoices.