[FIX] mail_push: iOS notifications
Before this commit, iOS notifications weren't displayed to the user.
So we had to adapt the FCM messages to make it work:

- 'priority':
We want to deliver notification instantly. So, we set it to 'high'.
- 'content_available':
Wake up the client app. On android, it wasn't necessary to precise
because data messages wake the app by default.
- 'notification':
iOS not supporting background push notification without notification key.
For more details about these keys, see:
https://firebase.google.com/docs/cloud-messaging/http-server-ref

Now all push message payload contains notification key with
basic details and data payload same for both android and ios.
Because of this, we also have to adapt the code of the android app.
See: odoo/mobile@92c9e7d

Due to this change, it's not possible to define the priority
of the Android notification when the app is inactive.
Before we only used data message and the 'onMessageReceived'
callback was used to trigger a notification where we could
set the priority.
But now, we use messages with both notification and data payload.
In this case, the notification is delivered to the device’s
system tray and we can't change the priority.
The user can change it himself depending of the android version:
Settings -> Notifications -> Odoo.

For more info about how to handling FCM messages, see:
https://firebase.google.com/docs/cloud-messaging/android/receive