[FIX] delivery_*: adapt delivery carrier test to package weight
As the picking has not been validated yet, the packages weight cannot be computed
based on the quant, but needs to be computed based on the move_lines of
the current pickig.

So we give the picking_id in the context of the package we want the
weight for, it is necessary to avoid taking into account mova lines of
all picking for a package and be overweighted. Since commit https://github.com/odoo/odoo/commit/96c3001f0eed0de53b8f5a11a371767a5d25337d