[FIX] pos_loyalty: get_price resale reward
In odoo/odoo@ac89783d the change for product price handling were not
propagated the use in pos_loyalty.

This caused a "Resale" reward to be applied as much as available because
we always got {available point * price reward product} equals 0 which
was lower than the order total amount.

This commit correct the issue by getting correctly the price (with
quantity 1 since reward product are not merged).

opw-783864
closes #1737